pyproject.toml
[project]
name = "amplifier-app-vision"
version = "0.1.0"
description = "Image analysis and vision capabilities powered by OpenAI GPT-4 Vision"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Microsoft MADE:Explorations Team"}
]
keywords = ["vision", "image-analysis", "gpt-4-vision", "ai", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
vision = "amplifier_app_vision.cli:cli"

[project.urls]
Homepage = "https://github.com/microsoft/amplifier-app-vision"
Repository = "https://github.com/microsoft/amplifier-app-vision"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_app_vision"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
